:rotating_light: fix templating / data / ssr

+++ packages/api-gateway/src/types/cv/duration.js
@@ -1,6 +1,6 @@
 const duration = `{
-  from: Int!
-  to: Int
+  from: Float!
+  to: Float
 }`
 
 module.exports = `
